@@ -6,17 +6,19 @@ publish_to: "none"
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
-  cupertino_icons: ^1.0.0
+  cupertino_icons: ^1.0.3
   flutter_audio_recorder:
-  path_provider:
-  file:
-  audioplayers:
+    git:
+      url: https://github.com/devaliakbar/flutter_audio_recorder.git
+  path_provider: ^2.0.2
+  file: ^6.1.2
+  audioplayers: ^0.19.0
 
 dev_dependencies:
   flutter_test:
