@@ -3,16 +3,16 @@ import 'package:flutter/material.dart';
 
 class AudioWavePlayerWidget extends StatelessWidget {
   AudioWavePlayerWidget({
-    @required this.audioWavePlayerController,
+    required this.audioWavePlayerController,
     this.height,
     this.activeBarColor = const Color(0xFF81B3C1),
     this.inActiveBarColor = Colors.black54,
   });
 
-  final AudioWavePlayerController audioWavePlayerController;
+  final AudioWavePlayerController? audioWavePlayerController;
 
   ///[height] is the height of the container
-  final double height;
+  final double? height;
 
   /// [activeBarColor] is the color of the bar
   final Color activeBarColor;
@@ -22,7 +22,7 @@ class AudioWavePlayerWidget extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    double _containerHeight = height;
+    double? _containerHeight = height;
     if (_containerHeight == null) {
       final double _blockSizeVertical =
           MediaQuery.of(context).size.height / 100;
@@ -36,36 +36,36 @@ class AudioWavePlayerWidget extends StatelessWidget {
     final double _spacing = 0.3 * _blockSizeHorizontal;
 
     return StreamBuilder(
-      stream: audioWavePlayerController.barAnimationStream?.stream ?? null,
+      stream: audioWavePlayerController!.barAnimationStream?.stream ?? null,
       builder: (context, snapshot) {
         return Container(
           height: _containerHeight,
           child: ListView.builder(
             shrinkWrap: true,
             padding: EdgeInsets.all(0),
-            itemCount: audioWavePlayerController.audioWaves.length,
+            itemCount: audioWavePlayerController!.audioWaves!.length,
             scrollDirection: Axis.horizontal,
             itemBuilder: (BuildContext _, int index) {
               bool isPlayed = false;
               if (snapshot.data != null &&
-                  (audioWavePlayerController.audioWaveStatus ==
+                  (audioWavePlayerController!.audioWaveStatus ==
                           AudioWaveStatus.play ||
-                      audioWavePlayerController.audioWaveStatus ==
+                      audioWavePlayerController!.audioWaveStatus ==
                           AudioWaveStatus.pause)) {
-                final int indexUntilPlayed = snapshot.data;
+                final int indexUntilPlayed = snapshot.data as int;
                 if (index <= indexUntilPlayed) {
                   isPlayed = true;
                 }
               }
 
-              double bar = audioWavePlayerController.audioWaves[index] * 100;
+              double bar = audioWavePlayerController!.audioWaves![index] * 100;
 
               return Column(
                 mainAxisAlignment: MainAxisAlignment.center,
                 children: [
                   Container(
                     margin: EdgeInsets.only(right: _spacing),
-                    height: bar * _containerHeight / 100,
+                    height: bar * _containerHeight! / 100,
                     width: _barWidth,
                     decoration: BoxDecoration(
                       color: isPlayed ? activeBarColor : inActiveBarColor,
